отрефакторил окно фаворитов
