+++ project/src/components/favorite-list/favorite-item.tsx
@@ -1,0 +1,44 @@
+import { Link } from "react-router-dom";
+import Rating from "../rating/rating";
+import { Offer } from "../../types/offer";
+
+type FavoritesPlacesProps = {
+  offer: Offer;
+}
+
+function FavoriteItem({offer}: FavoritesPlacesProps) :JSX.Element {
+  const { id, isPremium, previewImage, price, isFavorite, rating, title, type, city } = offer;
+
+  return(
+    <article key={`${city}-${id}`} className="favorites__card place-card">
+      {isPremium ? <div className="place-card__mark"><span>Premium</span></div> : ''}
+      <div className="favorites__image-wrapper place-card__image-wrapper">
+        <Link to="#">
+          <img className="place-card__image" src={previewImage} width="150" height="110" alt={`${title}`}/>
+        </Link>
+      </div>
+      <div className="favorites__card-info place-card__info">
+        <div className="place-card__price-wrapper">
+          <div className="place-card__price">
+            <b className="place-card__price-value">&euro;{price}</b>
+            <span className="place-card__price-text">&#47;&nbsp;night</span>
+          </div>
+          <button className={`place-card__bookmark-button button ${isFavorite ? 'place-card__bookmark-button--active' : ''}`} type="button">
+            <svg className="place-card__bookmark-icon" width="18" height="19">
+              <use xlinkHref="#icon-bookmark"></use>
+            </svg>
+            <span className="visually-hidden">In bookmarks</span>
+          </button>
+        </div>
+        <div className="place-card__rating rating">
+          <Rating rating={rating}/>
+        </div>
+        <h2 className="place-card__name">
+          <Link to="#">{title}</Link>
+        </h2>
+        <p className="place-card__type">{type}</p>
+      </div>
+    </article>
+  );
+}
+export default FavoriteItem;
