+++ project/src/components/favorite-list/favorite-places.tsx
@@ -1,0 +1,21 @@
+import { Offer } from "../../types/offer";
+import { useAppSelector } from "../../hooks";
+import { getFavoriteOffers } from "../../store/main-data/main-data.selectors";
+import FavoriteItem from "./favorite-item";
+
+type FavoritePlacesProps = {
+  city: string;
+}
+
+function FavoritePlaces({city}: FavoritePlacesProps) :JSX.Element {
+  const favoriteOffers: Offer[] = useAppSelector(getFavoriteOffers);
+
+  return(
+    <div className="favorites__places">
+      {favoriteOffers
+        .filter(({ city: { name } }) => name === city)
+        .map((offer) => (<FavoriteItem offer={offer}/>))}
+    </div>
+  );
+}
+export default FavoritePlaces;
